test: isolate broken nested config workspace
