+++ packages/vscode-extension/__tests__/suite-jsconfig-broken-nested/index.ts
@@ -1,0 +1,1 @@
+export { run } from '../runSuite';
