+++ packages/vscode-extension/__tests__/suite-jsconfig/jsconfig.test.ts
@@ -327,135 +327,6 @@ export default [{
     );
   });
 
-  test('a newly discovered broken nested config keeps valid ancestor config active', async () => {
-    const root = getWorkspaceRoot();
-    const rootConfigPath = path.join(root, 'rslint.config.js');
-    const originalRootConfig = fs.readFileSync(rootConfigPath, 'utf8');
-    const nestedDir = path.join(root, 'broken-nested-config');
-    const nestedFilePath = path.join(nestedDir, 'index.ts');
-    const nestedConfigPath = path.join(nestedDir, 'rslint.config.js');
-    const attemptedLoadPath = path.join(nestedDir, 'config-load-attempted');
-    const postFailureFilePath = path.join(nestedDir, 'post-failure.ts');
-    const rootDoc = await openFixture('index.ts');
-    let nestedDoc: vscode.TextDocument | undefined;
-    let postFailureDoc: vscode.TextDocument | undefined;
-
-    const rootConfigWithMarker = `export default [{
-  files: ['**/*.ts'],
-  languageOptions: {
-    parserOptions: { projectService: false, project: ['./tsconfig.json'] },
-  },
-  rules: {
-    '@typescript-eslint/no-unsafe-member-access': 'warn',
-    'no-debugger': 'error',
-  },
-  plugins: ['@typescript-eslint'],
-}];
-`;
-
-    await withFailClosedCleanup(
-      async () => {
-        fs.mkdirSync(nestedDir, { recursive: true });
-        fs.writeFileSync(nestedFilePath, 'debugger;\n', 'utf8');
-        fs.writeFileSync(rootConfigPath, rootConfigWithMarker, 'utf8');
-
-        await vscode.window.showTextDocument(rootDoc);
-        await waitForDiagnostics(rootDoc, (diags) =>
-          diags.some(
-            (diagnostic) =>
-              diagnostic.message.includes('no-unsafe-member-access') &&
-              diagnostic.severity === vscode.DiagnosticSeverity.Warning,
-          ),
-        );
-        nestedDoc = await vscode.workspace.openTextDocument(nestedFilePath);
-        await vscode.window.showTextDocument(nestedDoc);
-        await waitForDiagnostics(nestedDoc, (diags) =>
-          diags.some(
-            (diagnostic) =>
-              diagnostic.message.includes('no-debugger') &&
-              diagnostic.severity === vscode.DiagnosticSeverity.Error,
-          ),
-        );
-        await closeTextEditor(nestedDoc);
-        nestedDoc = undefined;
-
-        fs.writeFileSync(
-          nestedConfigPath,
-          `import fs from 'node:fs';
-fs.writeFileSync(${JSON.stringify(attemptedLoadPath)}, 'attempted', 'utf8');
-throw new Error('intentional broken nested config');
-export default [];
-`,
-          'utf8',
-        );
-        await waitForFile(attemptedLoadPath);
-        assert.strictEqual(
-          fs.readFileSync(attemptedLoadPath, 'utf8'),
-          'attempted',
-          'The broken nested config must be evaluated before fallback is asserted',
-        );
-
-        // This URI does not exist until after the broken module has executed,
-        // so its diagnostics cannot be a stale snapshot from before the failed
-        // refresh. Its first lint must run after the blocking config transaction
-        // and resolve through the still-valid ancestor config.
-        fs.writeFileSync(postFailureFilePath, 'debugger;\n', 'utf8');
-        postFailureDoc =
-          await vscode.workspace.openTextDocument(postFailureFilePath);
-        await vscode.window.showTextDocument(postFailureDoc);
-        const postFailureDiagnostics = await waitForDiagnostics(
-          postFailureDoc,
-          (diags) =>
-            diags.some(
-              (diagnostic) =>
-                diagnostic.message.includes('no-debugger') &&
-                diagnostic.severity === vscode.DiagnosticSeverity.Error,
-            ),
-        );
-        assert.deepStrictEqual(
-          postFailureDiagnostics
-            .filter((diagnostic) => diagnostic.message.includes('no-debugger'))
-            .map((diagnostic) => diagnostic.severity),
-          [vscode.DiagnosticSeverity.Error],
-          'The valid ancestor must lint a file opened after the broken child was evaluated',
-        );
-      },
-      async () => {
-        const temporaryDocuments = [nestedDoc, postFailureDoc].filter(
-          (document): document is vscode.TextDocument => document !== undefined,
-        );
-        await withFailClosedCleanup(
-          async () => {
-            await Promise.all(
-              temporaryDocuments.map((document) => closeTextEditor(document)),
-            );
-          },
-          async () => {
-            // The temporary root config deliberately lowers this rule to a
-            // warning. Waiting for Error makes restoration an observable
-            // config-transaction barrier before the next test starts.
-            const rootRestored = waitForDiagnostics(rootDoc, (diags) =>
-              diags.some(
-                (diagnostic) =>
-                  diagnostic.message.includes('no-unsafe-member-access') &&
-                  diagnostic.severity === vscode.DiagnosticSeverity.Error,
-              ),
-            );
-            fs.writeFileSync(rootConfigPath, originalRootConfig, 'utf8');
-            fs.rmSync(nestedDir, { recursive: true, force: true });
-            assert.ok(
-              !fs.existsSync(nestedDir),
-              'Broken nested-config fixtures must be deleted during cleanup',
-            );
-            await rootRestored;
-          },
-          'Broken nested-config resource cleanup',
-        );
-      },
-      'Broken nested-config test',
-    );
-  });
-
   test('parent global ignores remove nested configs from the effective catalog', async () => {
     const root = getWorkspaceRoot();
     const rootConfigPath = path.join(root, 'rslint.config.js');
